Should of known that the value could be a float. Also fixed a typo.

+++ NeighbourhoodStatisticsExchange/src/nde2/types/delivery/DataSetItem.java
@@ -19,7 +19,7 @@ public class DataSetItem extends NDE2Result {
 	private Topic topic;
 	private Boundary boundary;
 	private Period period;
-	private int value;
+	private float value;
 
 	/**
 	 * @param topic
@@ -27,7 +27,8 @@ public class DataSetItem extends NDE2Result {
 	 * @param period
 	 * @param value
 	 */
-	public DataSetItem(Topic topic, Boundary boundary, Period period, int value) {
+	public DataSetItem(Topic topic, Boundary boundary, Period period,
+			float value) {
 		super(VALID_FOR_DAYS);
 		this.topic = topic;
 		this.boundary = boundary;
@@ -60,7 +61,7 @@ public class DataSetItem extends NDE2Result {
 	 * @return the value
 	 * @throws ValueNotAvailable
 	 */
-	public int getValue() throws ValueNotAvailable {
+	public float getValue() throws ValueNotAvailable {
 		if (value > Integer.MIN_VALUE)
 			return value;
 		else
